Safari-611.1.10/LayoutTests/accessibility/aria-roledescription-expected.txt

text
This tests that aria-roledescription works.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Role description: AXRoleDescription: Super Button.
Role description: AXRoleDescription: button.
Role description: AXRoleDescription: button.
Role description: AXRoleDescription: Super Button.
Role description: AXRoleDescription: button.
PASS successfullyParsed is true

TEST COMPLETE
